refactor: DRY getFieldName function
